added logic to include comments with post when retrieved by id tested with insomnia

+++ controllers/api/commentRoutes.js
@@ -6,6 +6,7 @@ const { log, info, warn, error } = require('@frenzie24/logger')
 router.post('/', withAuth, async (req, res) => {
     info('attempting new comment')
     try {
+
         const newComment = await Comment.create({
             ...req.body,
             user_id: req.session.user_id,
@@ -19,9 +20,16 @@ router.post('/', withAuth, async (req, res) => {
 });
 
 router.delete('/:id', withAuth, async (req, res) => {
-    const _id = req.params.id;
-    info(`attempting to delete Comment with id: ${_id}`)
+
     try {
+        const _id = Math.floor(req.params.id);
+        if (!Number.isInteger(_id)) {
+            warn(`Bad request: id invalid`);
+            res.status(400).json({ issue: 'id provided is invalid', solution: 'id needs to be an integer' });
+            return;
+        }
+
+        info(`attempting to delete Comment with id: ${_id}`)
         const commentData = await Comment.destroy({
             where: {
                 id: _id,
@@ -33,7 +41,7 @@ router.delete('/:id', withAuth, async (req, res) => {
             res.status(404).json({ message: 'No Comment found with this id!' });
             return;
         }
-        
+
         log('comment deleted', 'green');
         res.status(200).json(commentData);
     } catch (err) {
